Add php mime type
While PHP files don't have an official mime type adding these in will allow the package to identify various php extensions. The type used is one prefixed with text since PHP files should be human readable, this allows someone to identify a PHP file as plain text.

It's particularly useful to detect and then display the contents of a PHP file in a web app.